# code/episode-10-deploy-vercel/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Weather Pro",
    "version": "v1",
    "description": "Paid weather forecasts: current conditions and a 7-day outlook per city, with a free list of supported locations."
  },
  "paths": {
    "/forecast": {
      "get": {
        "operationId": "get-forecast",
        "summary": "Fetch current conditions and 7-day forecast for a location",
        "description": "",
        "tags": [
          "Forecast"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "type": "string",
              "description": "City to forecast. Defaults to San Francisco. See /locations for valid values."
            },
            "description": "City to forecast. Defaults to San Francisco. See /locations for valid values."
          }
        ],
        "responses": {
          "200": {
            "description": "Current conditions and 7-day forecast",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment required — pay through the gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequired"
                }
              }
            }
          },
          "404": {
            "description": "Requested location is not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnknownLocationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "get-health",
        "summary": "Check service health and availability status",
        "description": "",
        "tags": [
          "Health"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/locations": {
      "get": {
        "operationId": "get-locations",
        "summary": "List the locations supported for weather forecast lookups",
        "description": "",
        "tags": [
          "Locations"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationsResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Forecast"
    },
    {
      "name": "Health"
    },
    {
      "name": "Locations"
    }
  ],
  "servers": [
    {
      "url": "",
      "description": "API server"
    }
  ],
  "components": {
    "schemas": {
      "DailyForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "ISO date (YYYY-MM-DD)"
          },
          "high_c": {
            "type": "number",
            "description": "Daily high in Celsius"
          },
          "low_c": {
            "type": "number",
            "description": "Daily low in Celsius"
          },
          "condition": {
            "type": "string",
            "description": "Summary condition, e.g. \"sunny\""
          }
        },
        "required": [
          "date",
          "high_c",
          "low_c",
          "condition"
        ]
      },
      "ForecastResponse": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string",
            "description": "Canonical location name"
          },
          "generated_at": {
            "type": "string",
            "description": "ISO timestamp when the forecast was generated"
          },
          "current": {
            "type": "object",
            "properties": {
              "temp_c": {
                "type": "number",
                "description": "Current temperature in Celsius"
              },
              "condition": {
                "type": "string",
                "description": "Current condition"
              },
              "wind_kph": {
                "type": "number",
                "description": "Current wind speed in km/h"
              }
            },
            "required": [
              "temp_c",
              "condition",
              "wind_kph"
            ],
            "description": "Current conditions"
          },
          "forecast": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyForecast"
            },
            "description": "Seven-day daily forecast"
          }
        },
        "required": [
          "location",
          "generated_at",
          "current",
          "forecast"
        ],
        "description": "Current conditions and 7-day forecast"
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Service status, e.g. \"ok\""
          }
        },
        "required": [
          "status"
        ],
        "description": "Service health payload"
      },
      "LocationsResponse": {
        "type": "object",
        "properties": {
          "locations": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string",
                  "description": "Normalized location key"
                },
                "name": {
                  "type": "string",
                  "description": "Canonical display name"
                }
              },
              "required": [
                "key",
                "name"
              ]
            },
            "description": "Supported forecast locations"
          },
          "hint": {
            "type": "string",
            "description": "How to call the paid forecast endpoint"
          }
        },
        "required": [
          "locations",
          "hint"
        ],
        "description": "Discovery list of supported locations"
      },
      "PaymentRequired": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error code, e.g. \"payment_required\""
          },
          "message": {
            "type": "string",
            "description": "How to complete payment through the gateway"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "description": "Payment required — returned by the paid gateway"
      },
      "UnknownLocationResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation"
          },
          "available": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Supported location names"
          }
        },
        "required": [
          "error",
          "message",
          "available"
        ],
        "description": "Returned when the requested location is not supported"
      }
    },
    "responses": {}
  }
}